Fixing off by 1 error in ascertainedSitePatterns. Previous version caused incorrect ascertainment when there was exactly 1 include or exclude pattern type. This version fixes the issue. The interpretation of from and to arguments is inclusive and 1-base (i.e. in xml the counting starts from 1 not 0).

+++ src/dr/math/Binomial.java
@@ -37,10 +37,10 @@ package dr.math;
 public class Binomial
 {
 	//
-	// Public stuff
+	// Public stuff; lnGamma is used, so parameters can be doubles.
 	//
 	
-	public static double logChoose(int n, int k){
+	public static double logChoose(double n, double k){
 		return GammaFunction.lnGamma(n + 1.0) -	GammaFunction.lnGamma(k + 1.0) 
 			- GammaFunction.lnGamma(n - k + 1.0);
 	}
